fixed file structuring, ephmatch
