+++ src/components/Components.tsx
@@ -1,8 +1,10 @@
 // React imports
 import React from "react";
 import { ButtonProps } from "../lib/types";
+import { NavLink } from "react-router-dom";
+import "./stylesheets/StylizedLink.css";
 
-const Button = ({ children, ...other }: ButtonProps) => {
+export const Button = ({ children, ...other }: ButtonProps) => {
   return (
     <button type="button" {...other}>
       {children}
@@ -10,4 +12,31 @@ const Button = ({ children, ...other }: ButtonProps) => {
   );
 };
 
-export default Button;
+export const StylizedLink = ({
+  to,
+  title,
+  end,
+  children,
+}: {
+  to: string;
+  title?: string;
+  end?: boolean;
+  children?: React.ReactNode;
+}) => {
+  const applyActiveLinkStyling = ({
+    isActive,
+  }: {
+    isActive: boolean;
+  }): string => (isActive ? "active-stylized-link" : "");
+
+  return (
+    <NavLink
+      to={to}
+      className={applyActiveLinkStyling}
+      title={title}
+      end={end ?? false}
+    >
+      {children}
+    </NavLink>
+  );
+};
